[Refactor][Ops] Consolidate op files into one file per op family
- Merge gqa_decode, gqa_decode_paged, gqa_sliding_window, gqa_sliding_window_varlen into gqa.py
- Merge mha_decode, mha_decode_paged into mha.py
- Merge mhc_pre, mhc_post into ops/mhc.py (moved out of attention/)
- Merge gated_deltanet_recurrence into gated_deltanet.py
- Rename deepseek_dsa_decode.py -> deepseek_dsa.py
- Rename deepseek_mla_decode.py -> deepseek_mla.py
- Rename gla_recurrence.py -> gated_linear_attn.py
- Delete backward-compat shims: ops/op.py, kernels/kernel.py
- Update all imports and re-exports in __init__.py files
