--- tileops/kernels/kernel.py
@@ -1,2 +1,0 @@
-# Backward-compat shim — use tileops.kernels.kernel_base instead.
-from tileops.kernels.kernel_base import *  # noqa: F401,F403
