+++ tileops/ops/attention/__init__.py
@@ -1,21 +1,25 @@
-from .deepseek_dsa_decode import DeepSeekSparseAttentionDecodeWithKVCacheFwdOp
-from .deepseek_mla_decode import MultiHeadLatentAttentionDecodeWithKVCacheFwdOp
+from .deepseek_dsa import DeepSeekSparseAttentionDecodeWithKVCacheFwdOp
+from .deepseek_mla import MultiHeadLatentAttentionDecodeWithKVCacheFwdOp
 from .deepseek_nsa import (
     MeanPoolingForwardOp,
     NSACmpFwdVarlenOp,
     NSAFwdVarlenOp,
     NSATopkVarlenOp,
 )
-from .gqa import GroupedQueryAttentionBwdOp, GroupedQueryAttentionFwdOp
-from .gqa_decode import GroupedQueryAttentionDecodeWithKVCacheFwdOp
-from .gqa_decode_paged import GroupedQueryAttentionDecodePagedWithKVCacheFwdOp
-from .gqa_sliding_window import GqaSlidingWindowFwdOp
-from .gqa_sliding_window_varlen import GqaSlidingWindowVarlenFwdOp
-from .mha import MultiHeadAttentionBwdOp, MultiHeadAttentionFwdOp
-from .mha_decode import MultiHeadAttentionDecodeWithKVCacheFwdOp
-from .mha_decode_paged import MultiHeadAttentionDecodePagedWithKVCacheFwdOp
-from .mhc_post import MHCPostOp
-from .mhc_pre import MHCPreOp
+from .gqa import (
+    GqaSlidingWindowFwdOp,
+    GqaSlidingWindowVarlenFwdOp,
+    GroupedQueryAttentionBwdOp,
+    GroupedQueryAttentionDecodePagedWithKVCacheFwdOp,
+    GroupedQueryAttentionDecodeWithKVCacheFwdOp,
+    GroupedQueryAttentionFwdOp,
+)
+from .mha import (
+    MultiHeadAttentionBwdOp,
+    MultiHeadAttentionDecodePagedWithKVCacheFwdOp,
+    MultiHeadAttentionDecodeWithKVCacheFwdOp,
+    MultiHeadAttentionFwdOp,
+)
 
 __all__ = [
     "DeepSeekSparseAttentionDecodeWithKVCacheFwdOp",
@@ -25,8 +29,6 @@ __all__ = [
     "GroupedQueryAttentionDecodePagedWithKVCacheFwdOp",
     "GroupedQueryAttentionDecodeWithKVCacheFwdOp",
     "GroupedQueryAttentionFwdOp",
-    "MHCPostOp",
-    "MHCPreOp",
     "MeanPoolingForwardOp",
     "MultiHeadAttentionBwdOp",
     "MultiHeadAttentionDecodePagedWithKVCacheFwdOp",
