--- tileops/ops/attention/gqa_decode.py
@@ -1,50 +1,0 @@
-from typing import Dict, Optional
-
-import torch
-import torch.nn.functional as F
-
-from tileops.kernels.attention import GqaDecodeKernel
-from tileops.kernels.kernel_base import Kernel
-
-from ..op_base import Op
-
-__all__ = ["GroupedQueryAttentionDecodeWithKVCacheFwdOp"]
-
-
-class GroupedQueryAttentionDecodeWithKVCacheFwdOp(Op):
-    """Layout: BSHD"""
-
-    def __init__(self,
-                 batch: int,
-                 heads: int,
-                 heads_kv: int,
-                 seqlen_kv: int,
-                 dim: int,
-                 dtype: torch.dtype = torch.float16,
-                 kernel_map: Optional[Dict[str, Kernel]] = None,
-                 tune: bool = False) -> None:
-        self.batch = batch
-        self.heads = heads
-        self.heads_kv = heads_kv
-        self.seqlen_kv = seqlen_kv
-        self.dim = dim
-
-        self.dtype = dtype
-
-        self.dispatch_kernel(kernel_map)
-        self.kernel = self.kernel_map["gqa_decode_kernel"](
-            batch, heads, heads_kv, seqlen_kv, dim, self.dtype, tune=tune)
-
-    @property
-    def default_kernel_map(self) -> Dict[str, Kernel]:
-        return {"gqa_decode_kernel": GqaDecodeKernel}
-
-    def forward(self, q: torch.Tensor, k: torch.Tensor, v: torch.Tensor) -> torch.Tensor:
-        real_seqlen_kv = k.shape[1]
-        if real_seqlen_kv < self.seqlen_kv:
-            k = F.pad(
-                k, pad=(0, 0, 0, 0, 0, self.seqlen_kv - real_seqlen_kv), mode='constant', value=0)
-            v = F.pad(
-                v, pad=(0, 0, 0, 0, 0, self.seqlen_kv - real_seqlen_kv), mode='constant', value=0)
-
-        return self.kernel(q, k, v, real_seqlen_kv)
