--- tileops/ops/attention/gqa_decode_paged.py
@@ -1,46 +1,0 @@
-from typing import Dict, Optional
-
-import torch
-
-from tileops.kernels.attention.gqa_decode_paged import GqaDecodePagedKernel
-from tileops.kernels.kernel_base import Kernel
-
-from ..op_base import Op
-
-__all__ = ["GroupedQueryAttentionDecodePagedWithKVCacheFwdOp"]
-
-
-class GroupedQueryAttentionDecodePagedWithKVCacheFwdOp(Op):
-    """Paged GQA decode with dynamic KV cache. Layout: Q [batch, heads, dim] (BHD);
-    K, V physical cache [seqlen_kv, heads_kv, dim]; real_seqlen_kv [batch]; block_table [batch, num_pages].
-    """
-
-    def __init__(self,
-                 batch: int,
-                 heads: int,
-                 heads_kv: int,
-                 seqlen_kv: int,
-                 dim: int,
-                 page_size: int,
-                 dtype: torch.dtype = torch.float16,
-                 kernel_map: Optional[Dict[str, Kernel]] = None,
-                 tune: bool = False) -> None:
-        self.batch = batch
-        self.heads = heads
-        self.heads_kv = heads_kv
-        self.seqlen_kv = seqlen_kv
-        self.dim = dim
-        self.page_size = page_size
-        self.dtype = dtype
-
-        self.dispatch_kernel(kernel_map)
-        self.kernel = self.kernel_map["gqa_decode_paged_kernel"](
-            batch, heads, heads_kv, seqlen_kv, dim, page_size, self.dtype, tune=tune)
-
-    @property
-    def default_kernel_map(self) -> Dict[str, Kernel]:
-        return {"gqa_decode_paged_kernel": GqaDecodePagedKernel}
-
-    def forward(self, q: torch.Tensor, k: torch.Tensor, v: torch.Tensor,
-                real_seqlen_kv: torch.Tensor, block_table: torch.Tensor) -> torch.Tensor:
-        return self.kernel(q, k, v, real_seqlen_kv, block_table)
