+++ tileops/ops/attention/mha.py
@@ -1,12 +1,15 @@
 from typing import Dict, Optional
 
 import torch
+import torch.nn.functional as F
 
 from tileops.kernels.attention import (
     FlashAttnBwdPostprocessKernel,
     FlashAttnBwdPreprocessKernel,
     MhaBwdKernel,
     MhaBwdWgmmaPipelinedKernel,
+    MhaDecodeKernel,
+    MhaDecodePagedKernel,
     MhaFwdKernel,
     MhaFwdWgmmaPipelinedKernel,
 )
@@ -15,7 +18,12 @@ from tileops.utils import is_hopper
 
 from ..op_base import Op
 
-__all__ = ["MultiHeadAttentionFwdOp", "MultiHeadAttentionBwdOp"]
+__all__ = [
+    "MultiHeadAttentionFwdOp",
+    "MultiHeadAttentionBwdOp",
+    "MultiHeadAttentionDecodeWithKVCacheFwdOp",
+    "MultiHeadAttentionDecodePagedWithKVCacheFwdOp",
+]
 
 
 class MultiHeadAttentionFwdOp(Op):
@@ -99,3 +107,79 @@ class MultiHeadAttentionBwdOp(Op):
         dk, dv = self.kernel(q, k, v, do, lse, delta, dq)
         dq = dq.to(self.dtype) if is_hopper() else self.post_kernel(dq)
         return dq, dk, dv
+
+
+class MultiHeadAttentionDecodeWithKVCacheFwdOp(Op):
+    """Layout: BSHD"""
+
+    def __init__(self,
+                 batch: int,
+                 heads: int,
+                 seqlen_q: int,
+                 seqlen_kv: int,
+                 dim: int,
+                 dtype: torch.dtype = torch.float16,
+                 kernel_map: Optional[Dict[str, Kernel]] = None,
+                 tune: bool = False) -> None:
+        self.batch = batch
+        self.heads = heads
+        self.seqlen_q = seqlen_q
+        self.seqlen_kv = seqlen_kv
+        self.dim = dim
+
+        self.dtype = dtype
+
+        self.dispatch_kernel(kernel_map)
+        self.kernel = self.kernel_map["mha_decode_kernel"](
+            batch, heads, seqlen_q, seqlen_kv, dim, False, self.dtype, tune=tune)
+
+    @property
+    def default_kernel_map(self) -> Dict[str, Kernel]:
+        return {"mha_decode_kernel": MhaDecodeKernel}
+
+    def forward(self, q: torch.Tensor, k: torch.Tensor, v: torch.Tensor) -> torch.Tensor:
+        real_seqlen_kv = k.shape[1]
+        if real_seqlen_kv < self.seqlen_kv:
+            k = F.pad(
+                k, pad=(0, 0, 0, 0, 0, self.seqlen_kv - real_seqlen_kv), mode='constant', value=0)
+            v = F.pad(
+                v, pad=(0, 0, 0, 0, 0, self.seqlen_kv - real_seqlen_kv), mode='constant', value=0)
+        return self.kernel(q, k, v, real_seqlen_kv)
+
+
+class MultiHeadAttentionDecodePagedWithKVCacheFwdOp(Op):
+    """Paged MHA decode with dynamic KV cache. Layout: Q [batch, seqlen_q, heads, dim] (BSHD);
+    K, V physical cache [seqlen_kv, heads, dim]; real_seqlen_kv [batch]; block_table [batch, num_pages].
+    """
+
+    def __init__(self,
+                 batch: int,
+                 heads: int,
+                 seqlen_q: int,
+                 seqlen_kv: int,
+                 dim: int,
+                 page_size: int,
+                 is_causal: bool = False,
+                 dtype: torch.dtype = torch.float16,
+                 kernel_map: Optional[Dict[str, Kernel]] = None,
+                 tune: bool = False) -> None:
+        self.batch = batch
+        self.heads = heads
+        self.seqlen_q = seqlen_q
+        self.seqlen_kv = seqlen_kv
+        self.dim = dim
+        self.page_size = page_size
+        self.is_causal = is_causal
+        self.dtype = dtype
+
+        self.dispatch_kernel(kernel_map)
+        self.kernel = self.kernel_map["mha_decode_paged_kernel"](
+            batch, heads, seqlen_q, seqlen_kv, dim, page_size, is_causal, self.dtype, tune=tune)
+
+    @property
+    def default_kernel_map(self) -> Dict[str, Kernel]:
+        return {"mha_decode_paged_kernel": MhaDecodePagedKernel}
+
+    def forward(self, q: torch.Tensor, k: torch.Tensor, v: torch.Tensor,
+                real_seqlen_kv: torch.Tensor, block_table: torch.Tensor) -> torch.Tensor:
+        return self.kernel(q, k, v, real_seqlen_kv, block_table)
