--- tileops/ops/attention/mha_decode.py
@@ -1,49 +1,0 @@
-from typing import Dict, Optional
-
-import torch
-import torch.nn.functional as F
-
-from tileops.kernels.attention import MhaDecodeKernel
-from tileops.kernels.kernel_base import Kernel
-
-from ..op_base import Op
-
-__all__ = ["MultiHeadAttentionDecodeWithKVCacheFwdOp"]
-
-
-class MultiHeadAttentionDecodeWithKVCacheFwdOp(Op):
-    """Layout: BSHD"""
-
-    def __init__(self,
-                 batch: int,
-                 heads: int,
-                 seqlen_q: int,
-                 seqlen_kv: int,
-                 dim: int,
-                 dtype: torch.dtype = torch.float16,
-                 kernel_map: Optional[Dict[str, Kernel]] = None,
-                 tune: bool = False) -> None:
-        self.batch = batch
-        self.heads = heads
-        self.seqlen_q = seqlen_q
-        self.seqlen_kv = seqlen_kv
-        self.dim = dim
-
-        self.dtype = dtype
-
-        self.dispatch_kernel(kernel_map)
-        self.kernel = self.kernel_map["mha_decode_kernel"](
-            batch, heads, seqlen_q, seqlen_kv, dim, False, self.dtype, tune=tune)
-
-    @property
-    def default_kernel_map(self) -> Dict[str, Kernel]:
-        return {"mha_decode_kernel": MhaDecodeKernel}
-
-    def forward(self, q: torch.Tensor, k: torch.Tensor, v: torch.Tensor) -> torch.Tensor:
-        real_seqlen_kv = k.shape[1]
-        if real_seqlen_kv < self.seqlen_kv:
-            k = F.pad(
-                k, pad=(0, 0, 0, 0, 0, self.seqlen_kv - real_seqlen_kv), mode='constant', value=0)
-            v = F.pad(
-                v, pad=(0, 0, 0, 0, 0, self.seqlen_kv - real_seqlen_kv), mode='constant', value=0)
-        return self.kernel(q, k, v, real_seqlen_kv)
