--- tileops/ops/attention/mha_decode_paged.py
@@ -1,48 +1,0 @@
-from typing import Dict, Optional
-
-import torch
-
-from tileops.kernels.attention.mha_decode_paged import MhaDecodePagedKernel
-from tileops.kernels.kernel_base import Kernel
-
-from ..op_base import Op
-
-__all__ = ["MultiHeadAttentionDecodePagedWithKVCacheFwdOp"]
-
-
-class MultiHeadAttentionDecodePagedWithKVCacheFwdOp(Op):
-    """Paged MHA decode with dynamic KV cache. Layout: Q [batch, seqlen_q, heads, dim] (BSHD);
-    K, V physical cache [seqlen_kv, heads, dim]; real_seqlen_kv [batch]; block_table [batch, num_pages].
-    """
-
-    def __init__(self,
-                 batch: int,
-                 heads: int,
-                 seqlen_q: int,
-                 seqlen_kv: int,
-                 dim: int,
-                 page_size: int,
-                 is_causal: bool = False,
-                 dtype: torch.dtype = torch.float16,
-                 kernel_map: Optional[Dict[str, Kernel]] = None,
-                 tune: bool = False) -> None:
-        self.batch = batch
-        self.heads = heads
-        self.seqlen_q = seqlen_q
-        self.seqlen_kv = seqlen_kv
-        self.dim = dim
-        self.page_size = page_size
-        self.is_causal = is_causal
-        self.dtype = dtype
-
-        self.dispatch_kernel(kernel_map)
-        self.kernel = self.kernel_map["mha_decode_paged_kernel"](
-            batch, heads, seqlen_q, seqlen_kv, dim, page_size, is_causal, self.dtype, tune=tune)
-
-    @property
-    def default_kernel_map(self) -> Dict[str, Kernel]:
-        return {"mha_decode_paged_kernel": MhaDecodePagedKernel}
-
-    def forward(self, q: torch.Tensor, k: torch.Tensor, v: torch.Tensor,
-                real_seqlen_kv: torch.Tensor, block_table: torch.Tensor) -> torch.Tensor:
-        return self.kernel(q, k, v, real_seqlen_kv, block_table)
