--- tileops/ops/gated_deltanet_recurrence.py
@@ -1,76 +1,0 @@
-from typing import Dict, Optional, Tuple
-
-import torch
-
-from tileops.kernels.gated_deltanet_recurrence import (
-    GatedDeltaNetDecodeFP32Kernel,
-    GatedDeltaNetDecodeKernel,
-)
-from tileops.kernels.kernel_base import Kernel
-
-from .op_base import Op
-
-__all__ = ["GatedDeltaNetDecodeOp"]
-
-
-class GatedDeltaNetDecodeOp(Op):
-    """Gated DeltaNet decode (single-step recurrence).
-
-    Computes one step of the gated delta rule:
-        S_t = S_{t-1} (alpha_t (I - beta_t k_t k_t^T)) + beta_t v_t k_t^T
-        o_t = S_t q_t
-
-    Layout: BHD (batch, head, dim).
-    Supports float32, float16, and bfloat16 with fp32 accumulation.
-
-    For fp32 dtype, dispatches to a dedicated FP32 kernel that uses
-    element-wise matvec instead of T.gemm to avoid TF32 mantissa truncation.
-    """
-
-    def __init__(
-        self,
-        batch: int,
-        heads: int,
-        dim_k: int,
-        dim_v: int,
-        dtype: torch.dtype = torch.float32,
-        kernel_map: Optional[Dict[str, Kernel]] = None,
-        tune: bool = False,
-    ) -> None:
-        self.batch = batch
-        self.heads = heads
-        self.dim_k = dim_k
-        self.dim_v = dim_v
-        self.dtype = dtype
-
-        self.dispatch_kernel(kernel_map)
-
-        # Dispatch: fp32 → FP32 kernel (no TF32), fp16/bf16 → TC kernel
-        if dtype == torch.float32:
-            kernel_cls = self.kernel_map["GatedDeltaNetDecodeFP32Kernel"]
-        else:
-            kernel_cls = self.kernel_map["GatedDeltaNetDecodeKernel"]
-        kernel_dtype = Kernel.dtype_to_str(dtype)
-        self.kernel = kernel_cls(
-            batch, heads, dim_k, dim_v,
-            dtype=kernel_dtype,
-            tune=tune,
-        )
-
-    @property
-    def default_kernel_map(self) -> Dict[str, Kernel]:
-        return {
-            "GatedDeltaNetDecodeKernel": GatedDeltaNetDecodeKernel,
-            "GatedDeltaNetDecodeFP32Kernel": GatedDeltaNetDecodeFP32Kernel,
-        }
-
-    def forward(
-        self,
-        q: torch.Tensor,
-        k: torch.Tensor,
-        v: torch.Tensor,
-        g: torch.Tensor,
-        beta: torch.Tensor,
-        state: torch.Tensor,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        return self.kernel(q, k, v, g, beta, state)
