--- tileops/ops/op.py
@@ -1,2 +1,0 @@
-# Backward-compat shim — use tileops.ops.op_base instead.
-from tileops.ops.op_base import *  # noqa: F401,F403
